+++ engine/runtime-instrument-common/src/main/scala/org/enso/interpreter/instrument/command/Command.scala
@@ -1,21 +1,26 @@
 package org.enso.interpreter.instrument.command
 
-import org.enso.interpreter.instrument.execution.RuntimeContext
+import org.enso.interpreter.instrument.execution.{Completion, RuntimeContext}
 import org.enso.polyglot.runtime.Runtime.{Api, ApiResponse}
 import org.enso.polyglot.runtime.Runtime.Api.RequestId
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext}
 
 /** Base command trait that encapsulates a function request. Uses
   * [[RuntimeContext]] to perform a request.
   */
 abstract class Command(maybeRequestId: Option[RequestId]) {
 
+  type Result[_]
+
   /** Executes a request.
     *
     * @param ctx contains suppliers of services to perform a request
     */
-  def execute(implicit ctx: RuntimeContext, ec: ExecutionContext): Future[Unit]
+  def execute(implicit
+    ctx: RuntimeContext,
+    ec: ExecutionContext
+  ): Result[Completion]
 
   override def toString: String = this.getClass.getSimpleName