+++ engine/runtime-instrument-common/src/main/scala/org/enso/interpreter/instrument/command/DeserializeLibrarySuggestionsCmd.scala
@@ -13,10 +13,10 @@ import scala.concurrent.{ExecutionContext, Future}
 class DeserializeLibrarySuggestionsCmd(
   maybeRequestId: Option[Api.RequestId],
   request: Api.DeserializeLibrarySuggestions
-) extends Command(maybeRequestId) {
+) extends AsynchronousCommand(maybeRequestId) {
 
   /** @inheritdoc */
-  override def execute(implicit
+  override def executeAsynchronously(implicit
     ctx: RuntimeContext,
     ec: ExecutionContext
   ): Future[Unit] = {