+++ lib/rust/ensogl/core/src/display/scene.rs
@@ -310,16 +310,10 @@ pub struct Keyboard {
 }
 
 impl Keyboard {
-    pub fn new() -> Self {
+    pub fn new(target: &web::EventTarget) -> Self {
         let frp = enso_frp::io::keyboard::Keyboard::default();
-        let bindings = Rc::new(enso_frp::io::keyboard::DomBindings::new(&frp));
+        let bindings = Rc::new(enso_frp::io::keyboard::DomBindings::new(target, &frp));
         Self { frp, bindings }
-    }
-}
-
-impl Default for Keyboard {
-    fn default() -> Self {
-        Self::new()
     }
 }
 
@@ -348,6 +342,7 @@ impl Dom {
         root.set_style_or_warn("height", "100vh");
         root.set_style_or_warn("width", "100vw");
         root.set_style_or_warn("display", "block");
+        root.set_style_or_warn("outline", "none");
         let root = web::dom::WithKnownShape::new(&root);
         Self { root, layers }
     }
@@ -882,7 +877,7 @@ impl SceneData {
         let mouse =
             Mouse::new(&frp, &dom.root, &variables, &display_mode, &pointer_target_registry);
         let disable_context_menu = Rc::new(web::ignore_context_menu(&dom.root));
-        let keyboard = Keyboard::new();
+        let keyboard = Keyboard::new(&web::window);
         let network = &frp.network;
         let extensions = Extensions::default();
         let bg_color_var = style_sheet.var("application.background");